Unison.Auth.CredentialManager

Documentation

saveCredentials :: CredentialManager -> CodeserverId -> CodeserverCredentials -> IO ()

Saves credentials to the active profile.

data CredentialManager

A CredentialManager knows how to load, save, and cache credentials.

It's thread-safe and safe for use across multiple UCM clients.

Note: Currently the in-memory cache is _not_ updated if a different UCM updates the credentials file, however this shouldn't pose any problems, since auth will still be refreshed if we encounter any auth failures on requests.

globalCredentialManager :: CredentialManager

A global CredentialManager instance/singleton.

newCredentialManager :: Maybe FilePath -> IO CredentialManager

getCodeserverCredentials :: CredentialManager -> CodeserverId -> IO (Either CredentialFailure CodeserverCredentials)

getOrCreatePersonalKey :: CredentialManager -> IO PersonalPrivateKey

Fetches the user's personal key from the active profile, if it exists. Otherwise it creates a new personal key, saves it to the active profile, and returns it.

isExpired :: CodeserverCredentials -> IO Bool

Checks whether CodeserverCredentials are expired.