Unison.Auth.PersonalKey

Documentation

data PersonalPrivateKey

A JWK representing a personal key

encodePrivateKey :: PersonalPrivateKey -> Value

Encode the private JWK.

I left off a ToJSON instance because I want to be explicit about when we're encoding the private key.

data PersonalPublicKey

publicKey :: PersonalPrivateKey -> PersonalPublicKey

generatePersonalKey :: MonadIO m => m PersonalPrivateKey

personalKeyThumbprint :: PersonalPrivateKey -> KeyThumbprint

signWithPersonalKey :: MonadIO m => PersonalPrivateKey -> ByteString -> m (Either Error PersonalKeySignature)

Sign arbitrary bytes using a personal private key

>>> key <- generatePersonalKey
>>> let msg = "Hello, world!"
>>> signature <- fromRight (error "failed to sign") <$> signWithPersonalKey key msg
>>> verifyWithPersonalKey (publicKey key) msg signature
True

verifyWithPersonalKey :: MonadIO m => PersonalPublicKey -> ByteString -> PersonalKeySignature -> m Bool

Verify a signature made with a personal private key

newtype PersonalKeySignature

Constructors

PersonalKeySignature

Fields

unPersonalKeySignature :: ByteString